chore: move test utils
